Use TypeErasedSubscription in GenericSubscription
