+++ src/agnocastlib/include/agnocast/agnocast_callback_info.hpp
@@ -4,20 +4,10 @@
 #include "agnocast/agnocast_epoll_update_dispatcher.hpp"
 #include "agnocast/agnocast_smart_pointer.hpp"
 
-#include <rclcpp/serialized_message.hpp>
-
-#include <rosidl_runtime_c/message_type_support_struct.h>
-
-#include <cstddef>
 #include <cstdint>
 #include <memory>
 #include <mutex>
 #include <type_traits>
-
-namespace rcpputils
-{
-class SharedLibrary;
-}
 
 namespace agnocast
 {
@@ -26,14 +16,6 @@ namespace agnocast
 // atomic wrap-around (overflow back to 0) during concurrent fetch_add calls.
 constexpr uint32_t MAX_CALLBACK_INFO_ID_SAFETY_MARGIN = 1000;
 constexpr uint32_t MAX_CALLBACK_INFO_ID = UINT32_MAX - MAX_CALLBACK_INFO_ID_SAFETY_MARGIN;
-
-/// Bundles the dynamically loaded typesupport library with its message handle.
-/// Keeping both together ensures the handle is never outlived by its library.
-struct TypeSupportBundle
-{
-  std::shared_ptr<rcpputils::SharedLibrary> library;
-  const rosidl_message_type_support_t * handle{nullptr};
-};
 
 struct AgnocastExecutable;
 
@@ -57,26 +39,6 @@ public:
   const std::type_info & type() const override { return typeid(T); }
 
   agnocast::ipc_shared_ptr<T> && get() && { return std::move(ptr_); }
-};
-
-// Class for a generic message type used by GenericSubscription.
-class RawMessagePtr : public AnyObject
-{
-  // We use std::byte for ipc_shared_ptr because the exact message type is unknown
-  // at compile time in GenericSubscription.
-  //
-  // This is safe because this pointer purely references a subscribed message.
-  // It does not handle memory allocation or deallocation (no delete is called);
-  // it strictly manages the reference count and releases the reference to the
-  // kernel module. (See ipc_shared_ptr implementation for details).
-  agnocast::ipc_shared_ptr<std::byte> ptr_;
-
-public:
-  explicit RawMessagePtr(agnocast::ipc_shared_ptr<std::byte> p) : ptr_(std::move(p)) {}
-
-  const std::type_info & type() const override { return typeid(RawMessagePtr); }
-
-  agnocast::ipc_shared_ptr<std::byte> && get() && { return std::move(ptr_); }
 };
 
 // Type for type-erased callback function
@@ -126,72 +88,6 @@ TypeErasedCallback get_erased_callback(Func && callback)
   };
 }
 
-// Serialize raw message data into a SerializedMessage.
-// Returns false (and logs an error) on rmw_serialize failure.
-bool serialize_message(
-  const void * raw, const rosidl_message_type_support_t * type_support,
-  rclcpp::SerializedMessage & out);
-
-template <typename Func>
-TypeErasedCallback get_erased_generic_callback(Func && callback, TypeSupportBundle ts)
-{
-  using F = std::decay_t<Func>;
-  static_assert(
-    std::is_invocable_v<F, std::shared_ptr<rclcpp::SerializedMessage>> ||
-      std::is_invocable_v<F, std::unique_ptr<rclcpp::SerializedMessage>> ||
-      std::is_invocable_v<F, rclcpp::SerializedMessage &>,
-    "This callback type cannot be handled as a GenericCallback. "
-    "Callback must be invocable with one of the following arguments "
-    "(or any types implicitly convertible from them, e.g., const variants): "
-    "std::unique_ptr<rclcpp::SerializedMessage>, "
-    "std::shared_ptr<rclcpp::SerializedMessage>, or "
-    "rclcpp::SerializedMessage &.");
-
-  return [callback = std::forward<Func>(callback), ts = std::move(ts)](AnyObject && arg) mutable {
-    if (typeid(RawMessagePtr) != arg.type()) {
-      RCLCPP_ERROR(
-        logger, "Agnocast internal implementation error: bad allocation when callback is called");
-      close(agnocast_fd);
-      exit(EXIT_FAILURE);
-    }
-
-    agnocast::ipc_shared_ptr<std::byte> raw_ptr =
-      std::move(static_cast<RawMessagePtr &&>(arg)).get();
-    if (!raw_ptr) {
-      RCLCPP_ERROR(logger, "received a null raw message pointer; skipping");
-      return;
-    }
-
-    if constexpr (std::is_invocable_v<F, std::shared_ptr<rclcpp::SerializedMessage>>) {
-      auto serialized = std::make_shared<rclcpp::SerializedMessage>();
-      if (!serialize_message(raw_ptr.get(), ts.handle, *serialized)) {
-        return;
-      }
-      raw_ptr.reset();
-      callback(std::move(serialized));
-    } else if constexpr (std::is_invocable_v<F, std::unique_ptr<rclcpp::SerializedMessage>>) {
-      auto serialized = std::make_unique<rclcpp::SerializedMessage>();
-      if (!serialize_message(raw_ptr.get(), ts.handle, *serialized)) {
-        return;
-      }
-      raw_ptr.reset();
-      callback(std::move(serialized));
-    } else {
-      rclcpp::SerializedMessage serialized;
-      if (!serialize_message(raw_ptr.get(), ts.handle, serialized)) {
-        return;
-      }
-      raw_ptr.reset();
-      callback(serialized);
-    }
-  };
-}
-
-uint32_t register_erased_callback(
-  TypeErasedCallback callback, MessageCreator message_creator, const std::string & topic_name,
-  topic_local_id_t subscriber_id, bool is_transient_local, mqd_t mqdes,
-  rclcpp::CallbackGroup::SharedPtr callback_group);
-
 template <typename MessageT, typename Func>
 uint32_t register_callback(
   Func && callback, const std::string & topic_name, const topic_local_id_t subscriber_id,
@@ -214,14 +110,24 @@ uint32_t register_callback(
       static_cast<MessageT *>(ptr), topic_name, subscriber_id, entry_id));
   };
 
-  return register_erased_callback(
-    std::move(erased_callback), std::move(message_creator), topic_name, subscriber_id,
-    is_transient_local, mqdes, std::move(callback_group));
-}
+  uint32_t callback_info_id = allocate_callback_info_id();
 
-uint32_t register_generic_callback(
-  TypeErasedCallback callback, const std::string & topic_name, topic_local_id_t subscriber_id,
-  bool is_transient_local, mqd_t mqdes, rclcpp::CallbackGroup::SharedPtr callback_group);
+  {
+    std::lock_guard<std::mutex> lock(id2_callback_info_mtx);
+    id2_callback_info[callback_info_id] = CallbackInfo{
+      topic_name,
+      subscriber_id,
+      is_transient_local,
+      mqdes,
+      std::move(callback_group),
+      std::move(erased_callback),
+      std::move(message_creator)};
+  }
+
+  EpollUpdateDispatcher::get_instance().request_update_all();
+
+  return callback_info_id;
+}
 
 void receive_and_execute_message(
   uint32_t callback_info_id, pid_t my_pid, const CallbackInfo & callback_info,
